[components] For https://github.com/mozilla-mobile/android-components/issues/5267 - Migrate telemetry setting
This patch introduces a new "FennecSettingsMigrator" which will be responsible
for migrating all common Fennec - Fenix settings.

Starting with the telemetry opt-in status which in Fenix will be "true" if the
Firefox Health Report setting ws also true on Fennec, and "false" otherwise.
It may be possible that the Fenix install is a new one and not an upgrade from
Fennec in which case no setting will be migrated and the migration would be
considered as successful.

Co-authored-by: Grisha Kruglov <gkruglovmozilla.com>

UltraBlame original commit: 07069f1047e8074b7b65ffe2986e9574a8387481