[components] Merge https://github.com/mozilla-mobile/android-components/pull/5334
5334: Closes https://github.com/mozilla-mobile/android-components/issues/5329: FennecMigrator: Use startForegroundService(). r=grigoryk a=pocmo

Whenever the app is in the background while we start a migration - for example when using a widget or some other background action triggers an Application object being created - then we try to launch the migration service and we see a crash since that is not allowed. Since we are using a foreground service anyways we can just use `startForegroundService()` here.

Co-authored-by: Sebastian Kaspari <s.kasparigmail.com>

UltraBlame original commit: 80fac00a1bd7f94d6ea2df7ced98ff2a9e72cb6f