[components] Merge https://github.com/mozilla-mobile/android-components/pull/5223
5223: Update to latest app-services, including metrics from logins component r=grigoryk a=rfk

(Top-level bug comment edited for clarity)

This updates a-c to use the latest a-s release, which includes some new metrics from the logins component. There was a lot of back-end-forth about the implications of landing such a thing, which you can read about below. The resolution is that we have data-review+ for landing these metrics in Fenix and Lockwise, and any new consumers of the logins component will need to get data-review for logins-component metrics.



Co-authored-by: Ryan Kelly <rfkellymozilla.com>

UltraBlame original commit: 919d032b5903e6ae760ef1058b4d67f3e1122cc7