[components] Merge https://github.com/mozilla-mobile/android-components/pull/5576 https://github.com/mozilla-mobile/android-components/pull/5577 (part 1)
5576: Switch to reuseAccount=false for the FxA migration r=jonalmeida a=grigoryk

The other option is what we really want, but it currently doesn't work.

(same as https://github.com/mozilla-mobile/android-components/pull/5571 but w/ tests passing)



5577: Issue https://github.com/mozilla-mobile/android-components/issues/5460: Exclude built-in extensions from addon migration r=grigoryk a=csadilek

There's **no** bug here currently, as built-in extensions are not being returned by `listInstalled` anyway. That's because they are not yet using the new install API, and are registered on every startup instead.

However, if we ever switch to the new API and persist those extensions as well, then this would be a footgun and we could by mistake migrate/disable them.

Co-authored-by: Grisha Kruglov <gkruglovmozilla.com>
Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 8700399b33f726889a51a633b3715fd408689ef8