[components] Merge https://github.com/mozilla-mobile/android-components/pull/5850
5850: Closes https://github.com/mozilla-mobile/android-components/issues/5692: NotImplementedError: Octet length not supported r=grigoryk a=st3fan

This patch for https://github.com/mozilla-mobile/android-components/issues/5692   fixes a integer sign issue when pulling bytes with the high bit set out of a `ByteArray`. I also fixed a bug where 3-byte lengths were not correctly decoded. (Unclear how likely it is that we encounter a login item that is larger than 2^16 though.)

I was not sure how to write a test for the embedded `ByteArray.getEncodingLength` so I actually pulled that extension method out of the `FennecLoginsMigration` class. Let me know if there is an elegant solution for that.

Draft because I'll add a few more test cases with funny numbers.



Co-authored-by: Stefan Arentz <stefanarentz.ca>

UltraBlame original commit: 1c10305f1faa2561830e2502c98fbd09c50af0db