[components] Merge https://github.com/mozilla-mobile/android-components/pull/6111
6111: Closes https://github.com/mozilla-mobile/android-components/issues/1912: Call speculativeConnect from toolbar and awesomebar r=pocmo a=csadilek

This introduces calls to `Engine.speculativeConnect` from our toolbar and awesomebar.

In the toolbar, we now call `speculativeConnect` on successful autocompletion from our domain list or history (re-using the existing logic to discard stale results.)

In the awesomebar, we now call `speculativeConnect` for the highest ranked search suggestion. This worked quite well when I tested this. When typing a search term instead of a URL, `speculativeConnect` will be called with the corresponding search URL of the configured search engine. So, when hitting enter, the URL will have already been "pre-fetched."

This is "inspired by": 
- https://searchfox.org/mozilla-esr68/source/mobile/android/base/java/org/mozilla/gecko/home/BrowserSearch.java#522-525
- https://searchfox.org/mozilla-esr68/source/mobile/android/base/java/org/mozilla/gecko/home/BrowserSearch.java#623-627

pocmo I've decided to break this apart from the speculative gecko session mechanism, which I'll work on as a follow-up.


Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 45d63aec44b13ce73f3231dea4e9082ce20a960a