[components] Closes https://github.com/mozilla-mobile/android-components/issues/6290: Do not remove FxA push scope on logout
We're caching the `fxaPushScope` in memory and also saving it to our
prefs. When we log out, we delete the pref, but the in-memory isn't
removed.

We could always fetch from the prefs instead of using an in-memory
caching, however this could be a bit expensive to always read from prefs
since we need to do this check when receiving any push message.

An easier solution would be to never remove the scope in prefs once
generated. It is safe to persist the push scope across different
accounts since the combination of scope + UAID is unique per install.

UltraBlame original commit: 5dc9b7efcd99a297bca2d2130339d55b7667cc3f