[components] Merge https://github.com/mozilla-mobile/android-components/pull/6412
6412: Closes https://github.com/mozilla-mobile/android-components/issues/6367: Do not intercept app initiated load requests by default r=rocketsroger a=csadilek

Our interceptors in Fenix and R-B do not need to be invoked for load requests initiated by the application (e.g. the user loading a URL via the toolbar). With this, the interceptors can control whether or not they want to be invoked in that case.

This is a performance optimization, as we want to minimize the time we block GeckoView from initiating the page load (it has to wait for the result of `onLoadRequest`.)

Full discussion in: https://github.com/mozilla-mobile/fenix/issues/8782#issuecomment-603411945


Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 2d0ac94d549a7cfd399919f4c6037dd28ae1d044