[components] Merge https://github.com/mozilla-mobile/android-components/pull/6715
6715: Issue https://github.com/mozilla-mobile/android-components/issues/6331: Add isDefault bool to TopSiteEntity and APIs for flagging default top sites r=ekager,Amejia481 a=gabrielluong

Fixes https://github.com/mozilla-mobile/android-components/issues/6331. 

Amejia481 Just to give you some context about the changes in this PR:

In Fenix, we have added 3 default top sites (Pocket, Wikipedia, YouTube) along with telemetry. 
It is not enough to just have a hardcoded list of these 3 default top sites since we can't differentiate from the user added top sites and top sites migrated from fennec.

We want to add this isDefault boolean to each TopSiteEntity so that in telemetry we can easily just query if the top site is a default top site added by the application. On the front end, we can also sort all the isDefault top sites to the beginning when we are showing the list of top sites.

In this PR, we did the following:

- Add isDefault bool to TopSiteEntity
- Migrate the DB to account for this new property
- Migrate the 3 default top sites to have its isDefault bool set to true



Co-authored-by: Gabriel Luong <gabriel.luonggmail.com>
Co-authored-by: Arturo Mejia <arturomejiamarmolgmail.com>

UltraBlame original commit: 130e625bd6ca0b2654bae685eac964c6512f666f