[components] Merge https://github.com/mozilla-mobile/android-components/pull/6828
6828: Issue https://github.com/mozilla-mobile/android-components/issues/1705 - lib-crash: Add database storage for crashes. r=rocketsroger a=pocmo

This patch adds the database classes to store information about crashes on disk. In this state it is not being used yet. Landing this separately to continue working on the UI and the integration. I'm not 100% happy with that part yet and therefore want to continue iterating. However the database piece seems to be stable enough and probably won't change.

Co-authored-by: Sebastian Kaspari <s.kasparigmail.com>

UltraBlame original commit: 3661a828612335c0eb8f3586aba09ea61d8c402e