[components] Merge https://github.com/mozilla-mobile/android-components/pull/6885
6885: Issue https://github.com/mozilla-mobile/android-components/issues/6884: SessionManager.onTrimMemory(): Do not remove engine session state from sessions. r=csadilek a=pocmo

When trimming memory we want to save the state of an engine session, attach it to the holder and close the engine session. Right now we try to do this for all sessions and then "unlink" them. This causes the engine session state to be cleared from sessions that either did not have an engine session (e.g. restored sessions) or that have been trimmed before. We should only perform this action for sessions that actually have an engine session. All other sessions can be left as-is.


Co-authored-by: Sebastian Kaspari <s.kasparigmail.com>

UltraBlame original commit: f4ea1a71ac3499e78f5a531b74137edf8f46eccf