[components] Merge https://github.com/mozilla-mobile/android-components/pull/7347
7347: Closes https://github.com/mozilla-mobile/android-components/issues/7244: PlacesHistoryStorage crash in Sample Browser r=grigoryk a=csadilek

Removes the `cleanup` call as it permanently leaves the `Connection` unusable and we're not re-initializing or creating a new instance of it. It's not clear to us how this should be used right now.

Also makes all access to `api` thread-safe.

Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 271ce4d4b596b0349d6766a4e2d66cb7961fba02