[components] Merge https://github.com/mozilla-mobile/android-components/pull/7759
7759: Issue https://github.com/mozilla-mobile/android-components/issues/7753: EngineViewPresenter: Temporarily disable releasing of engine session. r=csadilek a=pocmo

That was a new behavior that was introduced with the migration to `BrowserStore/State` and, surprise!, it can cause a white page of death - even though the order of events look correct and make sense. Anyhow, more in https://github.com/mozilla-mobile/android-components/issues/7753. I think we should temporarily disable this and continue to investigate the actual cause of the white page issue.

Co-authored-by: Sebastian Kaspari <s.kasparigmail.com>

UltraBlame original commit: 5d1207117669585fc051a3f644321b6d20c658e5