[components] Closes https://github.com/mozilla-mobile/android-components/issues/7665: Add support for converting TrackingProtectionPolicy to ContentBlocking.Setting.
We added the changes needed for Fenix to apply the policy settings they
need to the GeckoRuntime directly.

Since we require GeckoView changes to support the fix, we will let the
fix ride the train. However, to avoid breaking the build variants in
Fenix, we add the GeckoEngine extensions that are needed for the beta
variant. This would not give us the full benefits of the perf fixes,
but it will avoid regressing ETP changes should we upgrade the engine
and forget to apply changes we need for Fenix's GeckoProvider in beta.

Co-authored-by: Arturo Mejia <arturomejiamarmolgmail.com>

UltraBlame original commit: 1bab0e2969a641b9d4c9ecd92cb66c0b329c30c8