[components] Merge https://github.com/mozilla-mobile/android-components/pull/8047
8047: Closes https://github.com/mozilla-mobile/android-components/issues/8046: Remove unneeded unload handlers to make bfcache work r=jonalmeida a=csadilek

Window unload handlers block the bfcache. The extension framework already automatically disconnects ports on unload so they aren't needed in the first place.

For the reader mode content script, we are now connecting and disconnecting on pageshow/pagehide to support readerable checks when navigating (back/forward) to cached pages. We don't need to do that for reader pages, as we opted out of the cache there.

For FxA there's no difference in behaviour. I've also tested navigating back and forth during the signup flow and it still works, as expected.

For p2p, I opted not to change behaviour, but I am not familiar enough how it would be affected not that the bfcache is active.




Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: c6ec4a205377f2aee6e1b02bbd02e3be7a4f03eb