[components] Record send-tab telemetry for android-components consumers.
application-services needs to be closely involved in creating the
data to be recorded in telemetry, which passes it back to Android Components
as a JSON string. SyncTelemetry.kt is the only thing that needs to know
what's in that JSON, so it pull it apart and calls the relevant glean
functions.

See also https://github.com/mozilla/application-services/pull/3308

UltraBlame original commit: 1f4c5de49556df293e714ba38f06c224b689b62d